move connection and generic operation methods from RODSConn to RC_connect and RC_operation traits
